+++ tools/RAiDER/dem.py
@@ -19,7 +19,7 @@ from RAiDER.utilFcns import rio_open
 
 def download_dem(
         ll_bounds: Union[tuple, List, np.ndarray]=None,
-        dem_path: Path=Path('warpedDEM.dem'),
+        dem_path: Path=Path('warpedDEM.rdr'),
         overwrite: bool=False,
         writeDEM: bool=False,
         buf: float=0.02,